MGDAPI-5731 refactor probes
