+++ pkg/products/observability/reconciler_test.go
@@ -19,7 +19,6 @@ import (
 	"github.com/integr8ly/integreatly-operator/pkg/resources/quota"
 	"github.com/integr8ly/integreatly-operator/test/utils"
 	operatorsv1alpha1 "github.com/operator-framework/operator-lifecycle-manager/pkg/api/apis/operators/v1alpha1"
-	v1 "github.com/prometheus-operator/prometheus-operator/pkg/apis/monitoring/v1"
 	observability "github.com/redhat-developer/observability-operator/v4/api/v1"
 	corev1 "k8s.io/api/core/v1"
 	k8serr "k8s.io/apimachinery/pkg/api/errors"
@@ -558,114 +557,6 @@ func setupRecorder() record.EventRecorder {
 
 func getLogger() l.Logger {
 	return l.NewLoggerWithContext(l.Fields{l.ProductLogContext: "Dummy Product"})
-}
-
-func TestCreatePrometheusProbe(t *testing.T) {
-	type args struct {
-		ctx     context.Context
-		client  k8sclient.Client
-		inst    *v1alpha1.RHMI
-		cfg     *config.Observability
-		name    string
-		module  string
-		targets v1.ProbeTargetStaticConfig
-	}
-	tests := []struct {
-		name    string
-		args    args
-		want    v1alpha1.StatusPhase
-		wantErr bool
-	}{
-		{
-			name: "retry if namespace field is empty",
-			args: args{
-				ctx: context.TODO(),
-				cfg: &config.Observability{},
-			},
-			want:    v1alpha1.PhaseInProgress,
-			wantErr: false,
-		},
-		{
-			name: "retry if the URL(s) is not yet known",
-			args: args{
-				ctx: context.TODO(),
-				cfg: &config.Observability{
-					Config: map[string]string{"NAMESPACE": "ns"},
-				},
-				targets: v1.ProbeTargetStaticConfig{},
-			},
-			want:    v1alpha1.PhaseInProgress,
-			wantErr: false,
-		},
-		{
-			name: "failed to create probe",
-			args: args{
-				ctx: context.TODO(),
-				cfg: &config.Observability{
-					Config: map[string]string{"NAMESPACE": "ns"},
-				},
-				targets: v1.ProbeTargetStaticConfig{
-					Targets: []string{"testUrl"},
-				},
-				inst: &v1alpha1.RHMI{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "rhoam",
-						Namespace: "ns",
-					},
-				},
-				client: &clientMock.SigsClientInterfaceMock{
-					GetFunc: func(ctx context.Context, key types.NamespacedName, obj k8sclient.Object) error {
-						return nil
-					},
-					UpdateFunc: func(ctx context.Context, obj k8sclient.Object, opts ...k8sclient.UpdateOption) error {
-						return genericError
-					},
-				},
-			},
-			want:    v1alpha1.PhaseFailed,
-			wantErr: true,
-		},
-		{
-			name: "success creating probe",
-			args: args{
-				ctx: context.TODO(),
-				cfg: &config.Observability{
-					Config: map[string]string{"NAMESPACE": "ns"},
-				},
-				targets: v1.ProbeTargetStaticConfig{
-					Targets: []string{"testUrl"},
-				},
-				inst: &v1alpha1.RHMI{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      "rhoam",
-						Namespace: "ns",
-					},
-				},
-				client: &clientMock.SigsClientInterfaceMock{
-					GetFunc: func(ctx context.Context, key types.NamespacedName, obj k8sclient.Object) error {
-						return nil
-					},
-					UpdateFunc: func(ctx context.Context, obj k8sclient.Object, opts ...k8sclient.UpdateOption) error {
-						return nil
-					},
-				},
-			},
-			want:    v1alpha1.PhaseCompleted,
-			wantErr: false,
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			got, err := CreatePrometheusProbe(tt.args.ctx, tt.args.client, tt.args.inst, tt.args.cfg, tt.args.name, tt.args.module, tt.args.targets)
-			if (err != nil) != tt.wantErr {
-				t.Errorf("CreatePrometheusProbe() error = %v, wantErr %v", err, tt.wantErr)
-				return
-			}
-			if got != tt.want {
-				t.Errorf("CreatePrometheusProbe() got = %v, want %v", got, tt.want)
-			}
-		})
-	}
 }
 
 func TestReconciler_deleteObservabilityCR(t *testing.T) {
